More robust test suite
