+++ src/integration/init_test.go
@@ -1,0 +1,11 @@
+package integration
+
+import (
+	"testing"
+	. "launchpad.net/gocheck"
+)
+
+// Hook up gocheck into the gotest runner.
+func Test(t *testing.T) {
+	TestingT(t)
+}
