+++ src/integration/multiple_servers_test.go
@@ -1,93 +1,33 @@
 package integration
 
 import (
-	"bytes"
 	"common"
-	"configuration"
 	"crypto/tls"
 	"encoding/json"
 	"fmt"
+	. "integration/helpers"
 	"io/ioutil"
-	. "launchpad.net/gocheck"
 	"net"
 	"net/http"
 	"net/url"
 	"os"
-	"path/filepath"
-	"syscall"
 	"time"
+	. "launchpad.net/gocheck"
 )
 
 type ServerSuite struct {
-	serverProcesses []*ServerProcess
+	serverProcesses []*Server
 }
 
-type ServerProcess struct {
-	p          *os.Process
-	configFile string
-	apiPort    int
+var _ = Suite(&ServerSuite{})
+
+func (self *ServerSuite) precreateShards(server *Server, c *C) {
+	self.createShards(server, int64(3600), "false", c)
+	self.createShards(server, int64(86400), "true", c)
+	server.WaitForServerToSync()
 }
 
-func NewServerProcess(configFile string, apiPort int, d time.Duration, c *C) *ServerProcess {
-	s := &ServerProcess{configFile: configFile, apiPort: apiPort}
-	err := s.Start()
-	c.Assert(err, IsNil)
-	if d > 0 {
-		time.Sleep(d)
-	}
-	return s
-}
-
-func (self *ServerProcess) doesWalExist() error {
-	config := configuration.LoadConfiguration(self.configFile)
-	_, err := os.Stat(filepath.Join(config.WalDir, "log.1"))
-	return err
-}
-
-func (self *ServerProcess) Start() error {
-	if self.p != nil {
-		return fmt.Errorf("Server is already running with pid %d", self.p.Pid)
-	}
-
-	dir, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-
-	root := filepath.Join(dir, "..", "..")
-	filename := filepath.Join(root, "daemon")
-	config := filepath.Join(root, "src/integration/", self.configFile)
-	p, err := os.StartProcess(filename, []string{filename, "-config", config}, &os.ProcAttr{
-		Dir:   root,
-		Env:   os.Environ(),
-		Files: []*os.File{os.Stdin, os.Stdout, os.Stderr},
-	})
-	if err != nil {
-		return err
-	}
-	self.p = p
-	time.Sleep(5 * time.Second)
-	return nil
-}
-
-func (self *ServerProcess) Stop() {
-	if self.p == nil {
-		return
-	}
-
-	self.p.Signal(syscall.SIGTERM)
-	self.p.Wait()
-	self.p = nil
-}
-
-func (self *ServerSuite) precreateShards(server *ServerProcess, c *C) {
-	time.Sleep(time.Second)
-	go self.createShards(server, int64(3600), "false", c)
-	go self.createShards(server, int64(86400), "true", c)
-	time.Sleep(3 * time.Second)
-}
-
-func (self ServerSuite) createShards(server *ServerProcess, bucketSize int64, longTerm string, c *C) {
+func (self ServerSuite) createShards(server *Server, bucketSize int64, longTerm string, c *C) {
 	serverCount := 3
 	nowBucket := time.Now().Unix() / bucketSize * bucketSize
 	startIndex := 0
@@ -112,152 +52,15 @@ func (self ServerSuite) createShards(server *ServerProcess, bucketSize int64, lo
 	}
 }
 
-func ResultsToSeriesCollection(results []*common.SerializedSeries) *SeriesCollection {
-	return &SeriesCollection{Members: results}
-}
-
-type SeriesCollection struct {
-	Members []*common.SerializedSeries
-}
-
-func (self *SeriesCollection) GetSeries(name string, c *C) *Series {
-	for _, s := range self.Members {
-		if s.Name == name {
-			return &Series{s}
-		}
-	}
-	c.Fatalf("Couldn't find series '%s' in:\n", name, self)
-	return nil
-}
-
-type Series struct {
-	*common.SerializedSeries
-}
-
-func (self *Series) GetValueForPointAndColumn(pointIndex int, columnName string, c *C) interface{} {
-	columnIndex := -1
-	for index, name := range self.Columns {
-		if name == columnName {
-			columnIndex = index
-		}
-	}
-	if columnIndex == -1 {
-		c.Errorf("Couldn't find column '%s' in series:\n", columnName, self)
-		return nil
-	}
-	if pointIndex > len(self.Points)-1 {
-		c.Errorf("Fewer than %d points in series '%s':\n", pointIndex+1, self.Name, self)
-	}
-	return self.Points[pointIndex][columnIndex]
-}
-
-type Point struct {
-	Values []interface{}
-}
-
-func (self *ServerProcess) Query(database, query string, onlyLocal bool, c *C) *SeriesCollection {
-	return self.QueryWithUsername(database, query, onlyLocal, c, "paul", "pass")
-}
-
-func (self *ServerProcess) QueryAsRoot(database, query string, onlyLocal bool, c *C) *SeriesCollection {
-	return self.QueryWithUsername(database, query, onlyLocal, c, "root", "root")
-}
-
-func (self *ServerProcess) GetResponse(database, query, username, password string, onlyLocal bool, c *C) *http.Response {
-	encodedQuery := url.QueryEscape(query)
-	fullUrl := fmt.Sprintf("http://localhost:%d/db/%s/series?u=%s&p=%s&q=%s", self.apiPort, database, username, password, encodedQuery)
-	if onlyLocal {
-		fullUrl = fullUrl + "&force_local=true"
-	}
-	resp, err := http.Get(fullUrl)
-	c.Assert(err, IsNil)
-	return resp
-}
-
-func (self *ServerProcess) GetErrorBody(database, query, username, password string, onlyLocal bool, c *C) (string, int) {
-	resp := self.GetResponse(database, query, username, password, onlyLocal, c)
-	defer resp.Body.Close()
-	c.Assert(resp.StatusCode, Not(Equals), http.StatusOK)
-	bytes, err := ioutil.ReadAll(resp.Body)
-	c.Assert(err, IsNil)
-	return string(bytes), resp.StatusCode
-}
-
-func (self *ServerProcess) QueryWithUsername(database, query string, onlyLocal bool, c *C, username, password string) *SeriesCollection {
-	resp := self.GetResponse(database, query, username, password, onlyLocal, c)
-	defer resp.Body.Close()
-	c.Assert(resp.StatusCode, Equals, http.StatusOK)
-	body, err := ioutil.ReadAll(resp.Body)
-	c.Assert(err, IsNil)
-	var js []*common.SerializedSeries
-	err = json.Unmarshal(body, &js)
-	if err != nil {
-		fmt.Println("NOT JSON: ", string(body))
-	}
-	c.Assert(err, IsNil)
-	return ResultsToSeriesCollection(js)
-}
-
-func (self *ServerProcess) VerifyForbiddenQuery(database, query string, onlyLocal bool, c *C, username, password string) string {
-	encodedQuery := url.QueryEscape(query)
-	fullUrl := fmt.Sprintf("http://localhost:%d/db/%s/series?u=%s&p=%s&q=%s", self.apiPort, database, username, password, encodedQuery)
-	if onlyLocal {
-		fullUrl = fullUrl + "&force_local=true"
-	}
-	resp, err := http.Get(fullUrl)
-	c.Assert(err, IsNil)
-	defer resp.Body.Close()
-	c.Assert(resp.StatusCode, Equals, http.StatusForbidden)
-	body, err := ioutil.ReadAll(resp.Body)
-	c.Assert(err, IsNil)
-
-	return string(body)
-}
-
-func (self *ServerProcess) Post(url, data string, c *C) *http.Response {
-	err := self.Request("POST", url, data, c)
-	time.Sleep(time.Millisecond * 100)
-	return err
-}
-
-func (self *ServerProcess) Delete(url, body string, c *C) *http.Response {
-	err := self.Request("DELETE", url, body, c)
-	time.Sleep(time.Millisecond * 100)
-	return err
-}
-
-func (self *ServerProcess) PostGetBody(url, data string, c *C) []byte {
-	resp := self.Request("POST", url, data, c)
-	body, err := ioutil.ReadAll(resp.Body)
-	c.Assert(err, IsNil)
-	return body
-}
-
-func (self *ServerProcess) Get(url string, c *C) []byte {
-	resp := self.Request("GET", url, "", c)
-	body, err := ioutil.ReadAll(resp.Body)
-	c.Assert(err, IsNil)
-	return body
-}
-
-func (self *ServerProcess) Request(method, url, data string, c *C) *http.Response {
-	fullUrl := fmt.Sprintf("http://localhost:%d%s", self.apiPort, url)
-	req, err := http.NewRequest(method, fullUrl, bytes.NewBufferString(data))
-	c.Assert(err, IsNil)
-	resp, err := http.DefaultClient.Do(req)
-	c.Assert(err, IsNil)
-	return resp
-}
-
-var _ = Suite(&ServerSuite{})
-
 func (self *ServerSuite) SetUpSuite(c *C) {
 	err := os.RemoveAll("/tmp/influxdb/test")
 	c.Assert(err, IsNil)
-	self.serverProcesses = []*ServerProcess{
-		NewServerProcess("test_config1.toml", 60500, time.Second, c),
-		NewServerProcess("test_config2.toml", 60506, time.Second, c),
-		NewServerProcess("test_config3.toml", 60510, time.Second, c)}
+	self.serverProcesses = []*Server{
+		NewServer("src/integration/test_config1.toml", c),
+		NewServer("src/integration/test_config2.toml", c),
+		NewServer("src/integration/test_config3.toml", c),
+	}
+	self.serverProcesses[0].SetSslOnly(true)
 	self.serverProcesses[0].Post("/db?u=root&p=root", "{\"name\":\"full_rep\", \"replicationFactor\":3}", c)
 	self.serverProcesses[0].Post("/db?u=root&p=root", "{\"name\":\"test_rep\", \"replicationFactor\":2}", c)
 	self.serverProcesses[0].Post("/db?u=root&p=root", "{\"name\":\"single_rep\", \"replicationFactor\":1}", c)
@@ -267,7 +70,11 @@ func (self *ServerSuite) SetUpSuite(c *C) {
 	self.serverProcesses[0].Post("/db/single_rep/users?u=root&p=root", "{\"name\":\"paul\", \"password\":\"pass\", \"isAdmin\": true}", c)
 	self.serverProcesses[0].Post("/db/test_cq/users?u=root&p=root", "{\"name\":\"paul\", \"password\":\"pass\", \"isAdmin\": true}", c)
 	self.serverProcesses[0].Post("/db/test_cq/users?u=root&p=root", "{\"name\":\"weakpaul\", \"password\":\"pass\", \"isAdmin\": false}", c)
-	time.Sleep(time.Second)
+	self.serverProcesses[0].Post("/db?u=root&p=root", `{"name": "drop_db", "replicationFactor": 3}`, c)
+	self.serverProcesses[0].Post("/db/drop_db/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToSync()
+	}
 	self.precreateShards(self.serverProcesses[0], c)
 }
 
@@ -275,247 +82,6 @@ func (self *ServerSuite) TearDownSuite(c *C) {
 	for _, s := range self.serverProcesses {
 		s.Stop()
 	}
-}
-
-func (self *ServerSuite) TestWriteAndGetPoint(c *C) {
-	data := `
-  [{
-    "points": [[23.0]],
-    "name": "test_write_and_get_point",
-    "columns": ["something"]
-  }]
-  `
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-
-	collection := self.serverProcesses[0].Query("test_rep", "select * from test_write_and_get_point", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("test_write_and_get_point", c)
-	c.Assert(series.Points, HasLen, 1)
-	c.Assert(series.GetValueForPointAndColumn(0, "something", c).(float64), Equals, float64(23))
-}
-
-func (self *ServerSuite) TestWhereQuery(c *C) {
-	data := `[{"points": [[4], [10], [5]], "name": "test_where_query", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	collection := self.serverProcesses[0].Query("test_rep", "select * from test_where_query where value < 6", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("test_where_query", c)
-	c.Assert(series.Points, HasLen, 2)
-	c.Assert(series.GetValueForPointAndColumn(0, "value", c).(float64), Equals, float64(5))
-	c.Assert(series.GetValueForPointAndColumn(1, "value", c).(float64), Equals, float64(4))
-}
-
-func (self *ServerSuite) TestCountQueryOnSingleShard(c *C) {
-	data := `[{"points": [[4], [10], [5]], "name": "test_count_query_single_shard", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	t := (time.Now().Unix() - 60) * 1000
-	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_count_query_single_shard", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	collection := self.serverProcesses[0].Query("test_rep", "select count(value) from test_count_query_single_shard group by time(1m)", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("test_count_query_single_shard", c)
-	c.Assert(series.Points, HasLen, 2)
-	c.Assert(series.GetValueForPointAndColumn(0, "count", c).(float64), Equals, float64(3))
-	c.Assert(series.GetValueForPointAndColumn(1, "count", c).(float64), Equals, float64(1))
-}
-
-func (self *ServerSuite) TestRealGroupBy(c *C) {
-	data := `[{
-    "name": "sku_orders_hto_306-87",
-    "columns": ["time", "order_id", "order_product_id"],
-    "points": [
-      [1394308403000, 25798, 90727],
-      [1394096471000, 25619, 89815],
-      [1394023654000, 25558, 89552],
-      [1394017974000, 25545, 89487],
-      [1394017169000, 25542, 89478],
-      [1393961559000, 25510, 89321],
-      [1393758667000, 25330, 88488],
-      [1393662424000, 25250, 88131],
-      [1393660284000, 25249, 88123],
-      [1393316427000, 24859, 86603],
-      [1392995009000, 24559, 85177],
-      [1392837085000, 24375, 84268],
-      [1392703645000, 24179, 83451],
-      [1392636655000, 24109, 83091],
-      [1392462519000, 23917, 82141],
-      [1392401809000, 23879, 81962],
-      [1392304093000, 23747, 81296],
-      [1392289143000, 23711, 81130],
-      [1392275989000, 23681, 80966],
-      [1392207218000, 23606, 80615],
-      [1392105143000, 23492, 80079],
-      [1391707758000, 23194, 78680],
-      [1391681584000, 23145, 78430],
-      [1391678674000, 23139, 78397],
-      [1391638809000, 23112, 83088],
-      [1391627599000, 23092, 78178],
-      [1391547699000, 22999, 77679],
-      [1391460586000, 22922, 77317],
-      [1391416547000, 22853, 76974],
-      [1391353157000, 22814, 76789],
-      [1391343724000, 22804, 76737],
-      [1391283742000, 22766, 76543],
-      [1391181812000, 22653, 75973],
-      [1391072131000, 22459, 75032],
-      [1391060670000, 22437, 74936],
-      [1391018843000, 22399, 74754],
-      [1390988436000, 22327, 74374],
-      [1390940319000, 22283, 74160],
-      [1390891751000, 22119, 73365],
-      [1390890564000, 22112, 73323],
-      [1390890297000, 22107, 78521],
-      [1390890117000, 22104, 73291],
-      [1390888081000, 22096, 73252],
-      [1390887772000, 22095, 73249]
-    ]
-  }]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-
-	data = `[{
-    "name": "sku_orders_hto_306-67",
-    "columns": ["time", "order_id", "order_product_id"],
-    "points": [
-      [1394127830000, 25656, 89996],
-      [1394096295000, 25617, 89804],
-      [1394033181000, 25576, 89621],
-      [1394017169000, 25542, 89477],
-      [1393946314000, 25492, 89255],
-      [1393934255000, 25471, 89158],
-      [1393920481000, 25454, 89074],
-      [1393918574000, 25449, 89059],
-      [1393892147000, 25443, 89040],
-      [1393883034000, 25437, 89007],
-      [1393872507000, 25429, 88956],
-      [1393856915000, 25412, 88874],
-      [1393849643000, 25400, 88800],
-      [1393773937000, 25343, 88540],
-      [1393758667000, 25330, 88487],
-      [1393662424000, 25250, 88129],
-      [1393597853000, 25203, 87962],
-      [1393569612000, 25151, 87771],
-      [1393533783000, 25134, 87691],
-      [1393413072000, 24969, 87119],
-      [1393358360000, 24923, 86874],
-      [1393341285000, 24899, 86775],
-      [1393316427000, 24859, 86604],
-      [1393236191000, 24778, 86286],
-      [1393189696000, 24754, 86151],
-      [1393178560000, 24741, 86065],
-      [1393146070000, 24702, 85900],
-      [1393136028000, 24686, 85816],
-      [1393112374000, 24681, 85785],
-      [1393099370000, 24668, 85716],
-      [1393096719000, 24665, 85702],
-      [1393088205000, 24656, 85665],
-      [1393007664000, 24577, 85270],
-      [1392974360000, 24531, 85070],
-      [1392929913000, 24490, 84799],
-      [1392929102000, 24489, 84786],
-      [1392883165000, 24405, 84399],
-      [1392793278000, 24294, 83936],
-      [1392753876000, 24278, 83862],
-      [1392720329000, 24214, 83603],
-      [1392714746000, 24203, 83555],
-      [1392708003000, 24186, 83479],
-      [1392703645000, 24179, 83448],
-      [1392701245000, 24176, 83431],
-      [1392658814000, 24150, 83301],
-      [1392652348000, 24134, 83212],
-      [1392642305000, 24120, 83164],
-      [1392636655000, 24109, 83090],
-      [1392633742000, 24102, 83055],
-      [1392566655000, 24021, 82670],
-      [1392560788000, 24013, 82633],
-      [1392540563000, 23994, 82533],
-      [1392498475000, 23962, 82378],
-      [1392479295000, 23944, 82304],
-      [1392476505000, 23941, 82293],
-      [1392413768000, 23891, 82014],
-      [1392383113000, 23847, 81794],
-      [1392374154000, 23827, 81703],
-      [1392363800000, 23805, 81556],
-      [1392275989000, 23681, 80967],
-      [1392208228000, 23607, 80617],
-      [1392150829000, 23564, 80442],
-      [1392105143000, 23492, 80078],
-      [1392060844000, 23468, 79971],
-      [1391986799000, 23412, 79711],
-      [1391974922000, 23403, 79644],
-      [1391965155000, 23397, 79623],
-      [1391940963000, 23369, 79497],
-      [1391847849000, 23297, 79164],
-      [1391784730000, 23265, 79033],
-      [1391776948000, 23252, 78977],
-      [1391767550000, 23235, 78895],
-      [1391707758000, 23194, 78682],
-      [1391693189000, 23161, 78516],
-      [1391691391000, 23157, 78497],
-      [1391681584000, 23145, 78429],
-      [1391633408000, 23105, 78239],
-      [1391627599000, 23092, 78177],
-      [1391605518000, 23060, 78026],
-      [1391601531000, 23046, 77942],
-      [1391578346000, 23019, 77785],
-      [1391518604000, 22971, 77543],
-      [1391515497000, 22967, 77527],
-      [1391446215000, 22904, 77220],
-      [1391420126000, 22861, 77028],
-      [1391416547000, 22853, 76971],
-      [1391387621000, 22834, 76883],
-      [1391367769000, 22823, 76839],
-      [1391342864000, 22803, 76732],
-      [1391175187000, 22632, 75884],
-      [1391162331000, 22606, 75749],
-      [1391143666000, 22570, 75556],
-      [1391116400000, 22556, 75468],
-      [1391070045000, 22456, 75021],
-      [1391030959000, 22422, 74847],
-      [1390990512000, 22336, 74423],
-      [1390978045000, 22309, 74297],
-      [1390938734000, 22279, 77476],
-      [1390938108000, 22276, 74125],
-      [1390936641000, 22272, 74103],
-      [1390923498000, 22241, 73999],
-      [1390922359000, 22239, 73933],
-      [1390915811000, 22217, 73837],
-      [1390901530000, 22161, 73561],
-      [1390899150000, 22153, 73527],
-      [1390894824000, 22133, 73437],
-      [1390891700000, 22118, 73354],
-      [1390890564000, 22112, 73322],
-      [1390890117000, 22104, 73290],
-      [1390860722000, 22076, 77906],
-      [1390843479000, 22039, 72988]
-    ]
-  }]`
-
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	collection := self.serverProcesses[0].Query("test_rep", "select count(order_id) from sku_orders_hto_306-67 group by time(168h);", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("sku_orders_hto_306-67", c)
-	c.Assert(series.Points, HasLen, 7)
-	c.Assert(series.GetValueForPointAndColumn(0, "count", c).(float64), Equals, float64(2))
-	c.Assert(series.GetValueForPointAndColumn(1, "count", c).(float64), Equals, float64(17))
-	c.Assert(series.GetValueForPointAndColumn(2, "count", c).(float64), Equals, float64(18))
-	c.Assert(series.GetValueForPointAndColumn(3, "count", c).(float64), Equals, float64(23))
-	c.Assert(series.GetValueForPointAndColumn(4, "count", c).(float64), Equals, float64(16))
-	c.Assert(series.GetValueForPointAndColumn(5, "count", c).(float64), Equals, float64(18))
-	c.Assert(series.GetValueForPointAndColumn(6, "count", c).(float64), Equals, float64(17))
-}
-
-func (self *ServerSuite) TestGroupByDay(c *C) {
-	data := `[{"points": [[4], [10], [5]], "name": "test_group_by_day", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	t := (time.Now().Unix() - 86400) * 1000
-	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_group_by_day", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	collection := self.serverProcesses[0].Query("test_rep", "select count(value) from test_group_by_day group by time(1d)", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("test_group_by_day", c)
-	c.Assert(series.Points, HasLen, 2)
-	c.Assert(series.GetValueForPointAndColumn(0, "count", c).(float64), Equals, float64(3))
-	c.Assert(series.GetValueForPointAndColumn(1, "count", c).(float64), Equals, float64(1))
 }
 
 func (self *ServerSuite) TestGraphiteInterface(c *C) {
@@ -527,91 +93,21 @@ func (self *ServerSuite) TestGraphiteInterface(c *C) {
 
 	_, err = conn.Write([]byte(data))
 	c.Assert(err, IsNil)
+	conn.Close()
 
+	// there's no easy way to check whether the server started
+	// processing this request, unlike http requests which must return a
+	// status code
 	time.Sleep(time.Second)
+
+	self.serverProcesses[0].WaitForServerToSync()
 
 	collection := self.serverProcesses[0].QueryWithUsername("graphite_db", "select * from some_metric", false, c, "root", "root")
 	c.Assert(collection.Members, HasLen, 1)
 	series := collection.GetSeries("some_metric", c)
 	c.Assert(series.Points, HasLen, 2)
-	c.Assert(series.GetValueForPointAndColumn(0, "value", c).(float64), Equals, float64(200.5))
-	c.Assert(series.GetValueForPointAndColumn(1, "value", c).(float64), Equals, float64(100))
-}
-
-func (self *ServerSuite) TestLimitQueryOnSingleShard(c *C) {
-	data := `[{"points": [[4], [10], [5]], "name": "test_limit_query_single_shard", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	collection := self.serverProcesses[0].Query("test_rep", "select * from test_limit_query_single_shard limit 2", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("test_limit_query_single_shard", c)
-	c.Assert(series.Points, HasLen, 2)
-	c.Assert(series.GetValueForPointAndColumn(0, "value", c).(float64), Equals, float64(5))
-	c.Assert(series.GetValueForPointAndColumn(1, "value", c).(float64), Equals, float64(10))
-}
-
-func (self *ServerSuite) TestQueryAgainstMultipleShards(c *C) {
-	data := `[{"points": [[4], [10], [5]], "name": "test_query_against_multiple_shards", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	t := (time.Now().Unix() - 3600) * 1000
-	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_query_against_multiple_shards", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select count(value) from test_query_against_multiple_shards group by time(1h)", false, c)
-		c.Assert(collection.Members, HasLen, 1)
-		series := collection.GetSeries("test_query_against_multiple_shards", c)
-		c.Assert(series.Points, HasLen, 2)
-		c.Assert(series.GetValueForPointAndColumn(0, "count", c).(float64), Equals, float64(3))
-		c.Assert(series.GetValueForPointAndColumn(1, "count", c).(float64), Equals, float64(1))
-	}
-}
-
-func (self *ServerSuite) TestQueryAscendingAgainstMultipleShards(c *C) {
-	data := `[{"points": [[4], [10]], "name": "test_ascending_against_multiple_shards", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	t := (time.Now().Unix() - 3600) * 1000
-	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_ascending_against_multiple_shards", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select * from test_ascending_against_multiple_shards order asc", false, c)
-		series := collection.GetSeries("test_ascending_against_multiple_shards", c)
-		c.Assert(series.Points, HasLen, 3)
-		c.Assert(series.GetValueForPointAndColumn(0, "value", c).(float64), Equals, float64(2))
-		c.Assert(series.GetValueForPointAndColumn(1, "value", c).(float64), Equals, float64(4))
-		c.Assert(series.GetValueForPointAndColumn(2, "value", c).(float64), Equals, float64(10))
-	}
-}
-
-func (self *ServerSuite) TestBigGroupByQueryAgainstMultipleShards(c *C) {
-	data := `[{"points": [[4], [10]], "name": "test_multiple_shards_big_group_by", "columns": ["value"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	t := (time.Now().Unix() - 3600*2) * 1000
-	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_multiple_shards_big_group_by", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	time.Sleep(time.Second)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select count(value) from test_multiple_shards_big_group_by group by time(30d)", false, c)
-		series := collection.GetSeries("test_multiple_shards_big_group_by", c)
-		c.Assert(series.Points, HasLen, 1)
-		c.Assert(series.GetValueForPointAndColumn(0, "count", c).(float64), Equals, float64(3))
-	}
-}
-
-func (self *ServerSuite) TestWriteSplitToMultipleShards(c *C) {
-	data := `[
-		{"points": [[4], [10]], "name": "test_write_multiple_shards", "columns": ["value"]},
-		{"points": [["asdf"]], "name": "Test_write_multiple_shards", "columns": ["thing"]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select count(value) from test_write_multiple_shards", false, c)
-		series := collection.GetSeries("test_write_multiple_shards", c)
-		c.Assert(series.Points, HasLen, 1)
-		c.Assert(series.GetValueForPointAndColumn(0, "count", c).(float64), Equals, float64(2))
-
-		collection = s.Query("test_rep", "select * from Test_write_multiple_shards", false, c)
-		series = collection.GetSeries("Test_write_multiple_shards", c)
-		c.Assert(series.Points, HasLen, 1)
-		c.Assert(series.GetValueForPointAndColumn(0, "thing", c).(string), Equals, "asdf")
-	}
+	c.Assert(series.GetValueForPointAndColumn(0, "value", c), Equals, 200.5)
+	c.Assert(series.GetValueForPointAndColumn(1, "value", c), Equals, 100.0)
 }
 
 func (self *ServerSuite) TestRestartAfterCompaction(c *C) {
@@ -623,6 +119,7 @@ func (self *ServerSuite) TestRestartAfterCompaction(c *C) {
   }]
   `
 	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
+	self.serverProcesses[0].WaitForServerToSync()
 
 	collection := self.serverProcesses[0].Query("test_rep", "select * from test_restart_after_compaction", false, c)
 	c.Assert(collection.Members, HasLen, 1)
@@ -631,9 +128,8 @@ func (self *ServerSuite) TestRestartAfterCompaction(c *C) {
 	resp := self.serverProcesses[0].Post("/raft/force_compaction?u=root&p=root", "", c)
 	c.Assert(resp.StatusCode, Equals, http.StatusOK)
 	self.serverProcesses[0].Stop()
-	time.Sleep(time.Second)
 	self.serverProcesses[0].Start()
-	time.Sleep(time.Second * 3)
+	self.serverProcesses[0].WaitForServerToStart()
 
 	collection = self.serverProcesses[0].Query("test_rep", "select * from test_restart_after_compaction", false, c)
 	c.Assert(collection.Members, HasLen, 1)
@@ -641,7 +137,7 @@ func (self *ServerSuite) TestRestartAfterCompaction(c *C) {
 	c.Assert(series.Points, HasLen, 1)
 }
 
-func (self *ServerSuite) TestEntireClusterRestartAfterCompaction(c *C) {
+func (self *ServerSuite) TestEntireClusterReStartAfterCompaction(c *C) {
 	for i := 0; i < 3; i++ {
 		resp := self.serverProcesses[i].Post("/raft/force_compaction?u=root&p=root", "", c)
 		c.Assert(resp.StatusCode, Equals, http.StatusOK)
@@ -651,21 +147,17 @@ func (self *ServerSuite) TestEntireClusterRestartAfterCompaction(c *C) {
 		self.serverProcesses[i].Stop()
 	}
 
-	time.Sleep(3 * time.Second)
-
 	for i := 0; i < 3; i++ {
 		self.serverProcesses[i].Start()
 	}
 
-	time.Sleep(time.Second * 3)
-
 	for i := 0; i < 3; i++ {
-		self.serverProcesses[i].Get("/ping", c)
+		self.serverProcesses[i].WaitForServerToStart()
 	}
 }
 
 // For issue #140 https://github.com/influxdb/influxdb/issues/140
-func (self *ServerSuite) TestRestartServers(c *C) {
+func (self *ServerSuite) TestReStartServers(c *C) {
 	data := `
   [{
     "points": [[1]],
@@ -683,82 +175,19 @@ func (self *ServerSuite) TestRestartServers(c *C) {
 	for _, s := range self.serverProcesses {
 		s.Stop()
 	}
-	time.Sleep(time.Second)
 
-	err := self.serverProcesses[0].Start()
-	c.Assert(err, IsNil)
-	time.Sleep(time.Second)
-	err = self.serverProcesses[1].Start()
-	c.Assert(err, IsNil)
-	err = self.serverProcesses[2].Start()
-	time.Sleep(time.Second * 5)
+	for _, s := range self.serverProcesses {
+		c.Assert(s.Start(), IsNil)
+	}
+
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToStart()
+	}
 
 	collection = self.serverProcesses[0].Query("test_rep", "select * from test_restart", false, c)
 	c.Assert(collection.Members, HasLen, 1)
 	series = collection.GetSeries("test_restart", c)
 	c.Assert(series.Points, HasLen, 1)
-}
-
-func (self *ServerSuite) TestWritingNullInCluster(c *C) {
-	data := `[{"name":"test_null_in_cluster","columns":["provider","city"],"points":[["foo", "bar"], [null, "baz"]]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-
-	collection := self.serverProcesses[0].Query("test_rep", "select * from test_null_in_cluster", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	c.Assert(collection.GetSeries("test_null_in_cluster", c).Points, HasLen, 2)
-}
-
-func (self *ServerSuite) TestCountDistinctWithNullValues(c *C) {
-	data := `[{"name":"test_null_with_distinct","columns":["column"],"points":[["value1"], [null], ["value2"], ["value1"], [null]]}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-
-	collection := self.serverProcesses[0].Query("test_rep", "select count(distinct(column)) from test_null_with_distinct group by time(1m)", false, c)
-	c.Assert(collection.Members, HasLen, 1)
-	series := collection.GetSeries("test_null_with_distinct", c)
-	c.Assert(series.GetValueForPointAndColumn(0, "count", c), Equals, float64(2))
-}
-
-// issue #147
-func (self *ServerSuite) TestExtraSequenceNumberColumns(c *C) {
-	data := `
-  [{
-    "points": [
-        ["foo", 1390852524, 1234]
-    ],
-    "name": "test_extra_sequence_number_column",
-    "columns": ["val_1", "time", "sequence_number"]
-  }]`
-	resp := self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass&time_precision=s", data, c)
-	c.Assert(resp.StatusCode, Equals, http.StatusOK)
-
-	time.Sleep(time.Second)
-
-	collection := self.serverProcesses[0].Query("test_rep", "select * from test_extra_sequence_number_column", false, c)
-	series := collection.GetSeries("test_extra_sequence_number_column", c)
-	c.Assert(series.Columns, HasLen, 3)
-	c.Assert(series.GetValueForPointAndColumn(0, "sequence_number", c), Equals, float64(1234))
-}
-
-// issue #206
-func (self *ServerSuite) TestUnicodeSupport(c *C) {
-	data := `
-  [{
-    "points": [
-        ["山田太郎", "中文", "⚑"]
-    ],
-    "name": "test_unicode",
-    "columns": ["val_1", "val_2", "val_3"]
-  }]`
-	resp := self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	c.Assert(resp.StatusCode, Equals, http.StatusOK)
-
-	time.Sleep(time.Second)
-
-	collection := self.serverProcesses[0].Query("test_rep", "select * from test_unicode", false, c)
-	series := collection.GetSeries("test_unicode", c)
-	c.Assert(series.GetValueForPointAndColumn(0, "val_1", c), Equals, "山田太郎")
-	c.Assert(series.GetValueForPointAndColumn(0, "val_2", c), Equals, "中文")
-	c.Assert(series.GetValueForPointAndColumn(0, "val_3", c), Equals, "⚑")
 }
 
 func (self *ServerSuite) TestSslSupport(c *C) {
@@ -804,7 +233,7 @@ func (self *ServerSuite) TestInvalidUserNameAndDbName(c *C) {
 func (self *ServerSuite) TestShouldNotResetRootsPassword(c *C) {
 	resp := self.serverProcesses[0].Post("/db/dummy_db/users?u=root&p=root", "{\"name\":\"root\", \"password\":\"pass\"}", c)
 	c.Assert(resp.StatusCode, Equals, http.StatusOK)
-	time.Sleep(time.Second)
+	self.serverProcesses[0].WaitForServerToSync()
 	resp = self.serverProcesses[0].Request("GET", "/db/dummy_db/authenticate?u=root&p=pass", "", c)
 	c.Assert(resp.StatusCode, Equals, http.StatusOK)
 	resp = self.serverProcesses[0].Request("GET", "/cluster_admins/authenticate?u=root&p=root", "", c)
@@ -821,6 +250,7 @@ func (self *ServerSuite) TestDeleteFullReplication(c *C) {
     "columns": ["val_1", "val_2"]
   }]`
 	self.serverProcesses[0].Post("/db/full_rep/series?u=paul&p=pass", data, c)
+	self.serverProcesses[0].WaitForServerToSync()
 	collection := self.serverProcesses[0].Query("full_rep", "select count(val_1) from test_delete_full_replication", true, c)
 	series := collection.GetSeries("test_delete_full_replication", c)
 	c.Assert(series.GetValueForPointAndColumn(0, "count", c), Equals, float64(1))
@@ -842,6 +272,7 @@ func (self *ServerSuite) TestDeleteReplication(c *C) {
     "columns": ["val_1", "val_2"]
   }]`
 	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
+	self.serverProcesses[0].WaitForServerToSync()
 	collection := self.serverProcesses[0].Query("test_rep", "select count(val_1) from test_delete_replication", false, c)
 	series := collection.GetSeries("test_delete_replication", c)
 	c.Assert(series.GetValueForPointAndColumn(0, "count", c), Equals, float64(1))
@@ -865,6 +296,7 @@ func (self *ServerSuite) TestDbAdminPermissionToDeleteData(c *C) {
     "columns": ["val_1", "val_2"]
   }]`
 	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
+	self.serverProcesses[0].WaitForServerToSync()
 	collection := self.serverProcesses[0].QueryAsRoot("test_rep", "select count(val_1) from test_delete_admin_permission", false, c)
 	series := collection.GetSeries("test_delete_admin_permission", c)
 	c.Assert(series.GetValueForPointAndColumn(0, "count", c), Equals, float64(1))
@@ -884,6 +316,7 @@ func (self *ServerSuite) TestClusterAdminPermissionToDeleteData(c *C) {
     "columns": ["val_1", "val_2"]
   }]`
 	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
+	self.serverProcesses[0].WaitForServerToSync()
 	collection := self.serverProcesses[0].QueryAsRoot("test_rep", "select count(val_1) from test_delete_admin_permission", false, c)
 	series := collection.GetSeries("test_delete_admin_permission", c)
 	c.Assert(series.GetValueForPointAndColumn(0, "count", c), Equals, float64(1))
@@ -893,67 +326,23 @@ func (self *ServerSuite) TestClusterAdminPermissionToDeleteData(c *C) {
 	c.Assert(collection.Members, HasLen, 0)
 }
 
-func (self *ServerSuite) TestListSeries(c *C) {
-	self.serverProcesses[0].Post("/db?u=root&p=root", `{"name": "list_series", "replicationFactor": 2}`, c)
-	self.serverProcesses[0].Post("/db/list_series/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
-	time.Sleep(time.Second)
-	data := `[{
-		"name": "cluster_query",
-		"columns": ["val1"],
-		"points": [[1]]
-		}]`
-	self.serverProcesses[0].Post("/db/list_series/series?u=paul&p=pass", data, c)
-	t := (time.Now().Unix() - 3600) * 1000
-	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "another_query", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/list_series/series?u=paul&p=pass", data, c)
-
-	time.Sleep(time.Second)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("list_series", "list series", false, c)
-		c.Assert(collection.Members, HasLen, 2)
-		s := collection.GetSeries("cluster_query", c)
-		c.Assert(s, NotNil)
-		s = collection.GetSeries("another_query", c)
-		c.Assert(s, NotNil)
-	}
-}
-
-func (self *ServerSuite) TestSelectingTimeColumn(c *C) {
-	self.serverProcesses[0].Post("/db?u=root&p=root", `{"name": "test_rep", "replicationFactor": 2}`, c)
-	self.serverProcesses[0].Post("/db/test_rep/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
-	time.Sleep(time.Second)
-	data := `[{
-		"name": "selecting_time_column",
-		"columns": ["val1"],
-		"points": [[1]]
-		}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select val1, time from selecting_time_column", false, c)
-		s := collection.GetSeries("selecting_time_column", c)
-		c.Assert(s.Columns, HasLen, 3)
-		c.Assert(s.Points, HasLen, 1)
-	}
-}
-
 func (self *ServerSuite) TestDropDatabase(c *C) {
-	self.serverProcesses[0].Post("/db?u=root&p=root", `{"name": "drop_db", "replicationFactor": 3}`, c)
-	self.serverProcesses[0].Post("/db/drop_db/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
 	data := `[{
 		"name": "cluster_query",
 		"columns": ["val1"],
 		"points": [[1]]
 		}]`
 	self.serverProcesses[0].Post("/db/drop_db/series?u=paul&p=pass", data, c)
-	time.Sleep(time.Second)
+	self.serverProcesses[0].WaitForServerToSync()
 	resp := self.serverProcesses[0].Delete("/db/drop_db?u=root&p=root", "", c)
 	c.Assert(resp.StatusCode, Equals, http.StatusNoContent)
-	time.Sleep(time.Second)
 	self.serverProcesses[0].Post("/db?u=root&p=root", `{"name": "drop_db", "replicationFactor": 3}`, c)
 	self.serverProcesses[0].Post("/db/drop_db/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
-	time.Sleep(time.Second)
 	for _, s := range self.serverProcesses {
-		fmt.Printf("Running query against: %d\n", s.apiPort)
+		s.WaitForServerToSync()
+	}
+	for _, s := range self.serverProcesses {
+		fmt.Printf("Running query against: %d\n", s.ApiPort())
 		collection := s.Query("drop_db", "select * from cluster_query", true, c)
 		c.Assert(collection.Members, HasLen, 0)
 	}
@@ -963,15 +352,18 @@ func (self *ServerSuite) TestDropSeries(c *C) {
 	for i := 0; i < 3; i++ {
 		self.serverProcesses[0].Post("/db?u=root&p=root", `{"name": "drop_series", "replicationFactor": 3}`, c)
 		self.serverProcesses[0].Post("/db/drop_series/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
+		for _, s := range self.serverProcesses {
+			s.WaitForServerToSync()
+		}
 		data := `[{
 		"name": "cluster_query.1",
 		"columns": ["val1"],
 		"points": [[1]]
 		}]`
 		self.serverProcesses[0].Post("/db/drop_series/series?u=paul&p=pass", data, c)
-		time.Sleep(time.Second)
+		self.serverProcesses[0].WaitForServerToSync()
 		for _, s := range self.serverProcesses {
-			fmt.Printf("Running query against: %d\n", s.apiPort)
+			fmt.Printf("Running query against: %d\n", s.ApiPort())
 			collection := s.Query("drop_series", "select * from cluster_query.1", true, c)
 			c.Assert(collection.Members, HasLen, 1)
 			series := collection.GetSeries("cluster_query.1", c)
@@ -990,9 +382,13 @@ func (self *ServerSuite) TestDropSeries(c *C) {
 			c.Assert(resp.StatusCode, Equals, http.StatusNoContent)
 			self.serverProcesses[0].Post("/db/drop_series/users?u=root&p=root", `{"name": "paul", "password": "pass"}`, c)
 		}
-		time.Sleep(time.Second)
+
 		for _, s := range self.serverProcesses {
-			fmt.Printf("Running query against: %d\n", s.apiPort)
+			s.WaitForServerToSync()
+		}
+
+		for _, s := range self.serverProcesses {
+			fmt.Printf("Running query against: %d\n", s.ApiPort())
 			collection := s.Query("drop_series", "select * from cluster_query.1", true, c)
 			c.Assert(collection.Members, HasLen, 0)
 		}
@@ -1012,20 +408,20 @@ func (self *ServerSuite) TestRelogging(c *C) {
 
 	self.serverProcesses[0].Post("/db/full_rep/series?u=paul&p=pass", data, c)
 
-	time.Sleep(time.Second) // wait for data to get replicated
+	self.serverProcesses[0].WaitForServerToSync()
 
 	self.serverProcesses[0].Query("full_rep", "delete from test_relogging", false, c)
 
-	time.Sleep(time.Second)
-
 	for _, server := range self.serverProcesses[1:] {
-		err := server.doesWalExist()
+		err := server.DoesWalExist()
 		c.Assert(os.IsNotExist(err), Equals, true)
 		server.Stop()
-		time.Sleep(time.Second)
 		server.Start()
-		time.Sleep(time.Second)
-		err = server.doesWalExist()
+	}
+
+	for _, server := range self.serverProcesses[1:] {
+		server.WaitForServerToStart()
+		err := server.DoesWalExist()
 		c.Assert(os.IsNotExist(err), Equals, true)
 	}
 }
@@ -1042,7 +438,7 @@ func (self *ServerSuite) TestFailureAndReplicationReplays(c *C) {
   }]`
 	self.serverProcesses[0].Post("/db/full_rep/series?u=paul&p=pass", data, c)
 
-	time.Sleep(time.Second) // wait for data to get replicated
+	self.serverProcesses[0].WaitForServerToSync()
 
 	for _, s := range self.serverProcesses {
 		collection := s.Query("full_rep", "select sum(val) from test_failure_replays;", true, c)
@@ -1051,7 +447,9 @@ func (self *ServerSuite) TestFailureAndReplicationReplays(c *C) {
 	}
 
 	self.serverProcesses[1].Stop()
+	// wait for the server to be marked down
 	time.Sleep(time.Second)
+
 	data = `
 	[{
 		"points": [[2]],
@@ -1073,10 +471,11 @@ func (self *ServerSuite) TestFailureAndReplicationReplays(c *C) {
 	}
 
 	self.serverProcesses[1].Start()
-	time.Sleep(2 * time.Second)
+	self.serverProcesses[1].WaitForServerToStart()
+	self.serverProcesses[0].WaitForServerToSync()
 
 	for i := 0; i < 3; i++ {
-		// wait for the server to startup and the WAL to be synced
+		// wait for the server to Startup and the WAL to be synced
 		collection := self.serverProcesses[1].Query("full_rep", "select sum(val) from test_failure_replays;", true, c)
 		series := collection.GetSeries("test_failure_replays", c)
 		if series.GetValueForPointAndColumn(0, "sum", c).(float64) == 3 {
@@ -1084,71 +483,6 @@ func (self *ServerSuite) TestFailureAndReplicationReplays(c *C) {
 		}
 	}
 	c.Error("write didn't replay properly")
-}
-
-// func (self *ServerSuite) TestFailureAndDeleteReplays(c *C) {
-// 	data := `
-//   [{
-//     "points": [
-//         [1]
-//     ],
-//     "name": "test_failure_delete_replays",
-//     "columns": ["val"]
-//   }]`
-// 	self.serverProcesses[0].Post("/db/full_rep/series?u=paul&p=pass", data, c)
-// 	fmt.Println("TEST: posted failure")
-// 	for _, s := range self.serverProcesses {
-// 		collection := s.Query("full_rep", "select val from test_failure_delete_replays", true, c)
-// 		series := collection.GetSeries("test_failure_delete_replays", c)
-// 		c.Assert(series.Points, HasLen, 1)
-// 	}
-// 	fmt.Println("TEST: did queries, now stopping")
-// 	self.serverProcesses[1].Stop()
-// 	fmt.Println("TEST: running delete query on server 0")
-// 	self.serverProcesses[0].Query("full_rep", "delete from test_failure_delete_replays", false, c)
-// 	fmt.Println("TEST: done!")
-// 	time.Sleep(time.Second)
-// 	for i, s := range self.serverProcesses {
-// 		if i == 1 {
-// 			continue
-// 		} else {
-// 			fmt.Println("TEST: running sum query on server ", i)
-// 			collection := s.Query("full_rep", "select sum(val) from test_failure_delete_replays;", true, c)
-
-// 			c.Assert(collection.Members, HasLen, 0)
-// 		}
-// 	}
-
-// 	fmt.Println("TEST: starting server")
-// 	self.serverProcesses[1].Start()
-// 	time.Sleep(2 * time.Second)
-// 	sum := 0
-// 	for i := 0; i < 3; i++ {
-// 		fmt.Println("TEST: running sum query on server")
-// 		collection := self.serverProcesses[1].Query("full_rep", "select sum(val) from test_failure_delete_replays;", true, c)
-// 		sum += len(collection.Members)
-// 	}
-
-// 	c.Assert(sum, Equals, 0)
-// }
-
-// For issue #130 https://github.com/influxdb/influxdb/issues/130
-func (self *ServerSuite) TestColumnNamesReturnInDistributedQuery(c *C) {
-	data := `[{
-		"name": "cluster_query_with_columns",
-		"columns": ["col1"],
-		"points": [[1], [2]]
-		}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select * from cluster_query_with_columns", false, c)
-		series := collection.GetSeries("cluster_query_with_columns", c)
-		set := map[float64]bool{}
-		for idx, _ := range series.Points {
-			set[series.GetValueForPointAndColumn(idx, "col1", c).(float64)] = true
-		}
-		c.Assert(set, DeepEquals, map[float64]bool{1: true, 2: true})
-	}
 }
 
 func generateHttpApiSeries(name string, n int) *common.SerializedSeries {
@@ -1165,61 +499,6 @@ func generateHttpApiSeries(name string, n int) *common.SerializedSeries {
 	}
 }
 
-func (self *ServerSuite) TestLimitWithRegex(c *C) {
-	// run the test once with less than POINT_BATCH_SIZE points and once
-	// with more than POINT_BATCH_SIZE points
-	for _, numberOfPoints := range []int{100, 1000} {
-		for i := 0; i < 100; i++ {
-			name := fmt.Sprintf("limit_with_regex_%d_%d", numberOfPoints, i)
-			series := generateHttpApiSeries(name, numberOfPoints)
-			bytes, err := json.Marshal([]*common.SerializedSeries{series})
-			c.Assert(err, IsNil)
-			resp := self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", string(bytes), c)
-			defer resp.Body.Close()
-			c.Assert(resp.StatusCode, Equals, http.StatusOK)
-		}
-		time.Sleep(time.Second * 2)
-		query := fmt.Sprintf("select * from /.*limit_with_regex_%d.*/ limit 1", numberOfPoints)
-		collection := self.serverProcesses[0].Query("test_rep", query, false, c)
-		// make sure all series get back 1 point only
-		for i := 0; i < 100; i++ {
-			table := fmt.Sprintf("limit_with_regex_%d_%d", numberOfPoints, i)
-			series := collection.GetSeries(table, c)
-			c.Assert(series.SerializedSeries.Points, HasLen, 1)
-			c.Assert(series.GetValueForPointAndColumn(0, "value", c), Equals, float64(numberOfPoints-1))
-		}
-	}
-}
-
-// For issue #131 https://github.com/influxdb/influxdb/issues/131
-func (self *ServerSuite) TestSelectFromRegexInCluster(c *C) {
-	data := `[{
-		"name": "cluster_regex_query",
-		"columns": ["col1", "col2"],
-		"points": [[1, "foo"], [23, "bar"]]
-		},{
-			"name": "cluster_regex_query_number2",
-			"columns": ["blah"],
-			"points": [[true]]
-		},{
-			"name": "Cluster_regex_query_3",
-			"columns": ["foobar"],
-			"points": [["asdf"]]
-			}]`
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
-	time.Sleep(time.Second * 2)
-	for _, s := range self.serverProcesses {
-		collection := s.Query("test_rep", "select * from /.*/ limit 1", false, c)
-		series := collection.GetSeries("cluster_regex_query", c)
-		c.Assert(series.GetValueForPointAndColumn(0, "col1", c), Equals, float64(23))
-		c.Assert(series.GetValueForPointAndColumn(0, "col2", c), Equals, "bar")
-		series = collection.GetSeries("cluster_regex_query_number2", c)
-		c.Assert(series.GetValueForPointAndColumn(0, "blah", c), Equals, true)
-		series = collection.GetSeries("Cluster_regex_query_3", c)
-		c.Assert(series.GetValueForPointAndColumn(0, "foobar", c), Equals, "asdf")
-	}
-}
-
 func (self *ServerSuite) TestContinuousQueryManagement(c *C) {
 	collection := self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
 	series := collection.GetSeries("continuous queries", c)
@@ -1229,12 +508,16 @@ func (self *ServerSuite) TestContinuousQueryManagement(c *C) {
 	c.Assert(response, Equals, "Insufficient permissions to create continuous query")
 
 	self.serverProcesses[0].QueryAsRoot("test_cq", "select * from foo into bar;", false, c)
+	self.serverProcesses[0].WaitForServerToSync()
 
 	collection = self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
 	series = collection.GetSeries("continuous queries", c)
 	c.Assert(series.Points, HasLen, 1)
 	c.Assert(series.GetValueForPointAndColumn(0, "id", c), Equals, float64(1))
 	c.Assert(series.GetValueForPointAndColumn(0, "query", c), Equals, "select * from foo into bar;")
+
+	// wait for the continuous query to run
+	time.Sleep(time.Second)
 
 	self.serverProcesses[0].QueryAsRoot("test_cq", "select * from quu into qux;", false, c)
 	collection = self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
@@ -1246,6 +529,8 @@ func (self *ServerSuite) TestContinuousQueryManagement(c *C) {
 	c.Assert(series.GetValueForPointAndColumn(1, "query", c), Equals, "select * from quu into qux;")
 
 	self.serverProcesses[0].QueryAsRoot("test_cq", "drop continuous query 1;", false, c)
+	// wait for the continuous query to be dropped
+	self.serverProcesses[0].WaitForServerToSync()
 
 	collection = self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
 	series = collection.GetSeries("continuous queries", c)
@@ -1265,6 +550,7 @@ func (self *ServerSuite) TestContinuousQueryFanoutOperations(c *C) {
 	defer self.serverProcesses[0].QueryAsRoot("test_cq", "drop continuous query 2;", false, c)
 	defer self.serverProcesses[0].QueryAsRoot("test_cq", "drop continuous query 3;", false, c)
 	defer self.serverProcesses[0].QueryAsRoot("test_cq", "drop continuous query 4;", false, c)
+	self.serverProcesses[0].WaitForServerToSync()
 	collection := self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
 	series := collection.GetSeries("continuous queries", c)
 	c.Assert(series.Points, HasLen, 4)
@@ -1276,6 +562,7 @@ func (self *ServerSuite) TestContinuousQueryFanoutOperations(c *C) {
   ]`
 
 	self.serverProcesses[0].Post("/db/test_cq/series?u=paul&p=pass", data, c)
+	self.serverProcesses[0].WaitForServerToSync()
 
 	collection = self.serverProcesses[0].Query("test_cq", "select * from s1;", false, c)
 	series = collection.GetSeries("s1", c)
@@ -1333,10 +620,15 @@ func (self *ServerSuite) TestContinuousQueryGroupByOperations(c *C) {
 
 	self.serverProcesses[0].Post("/db/test_cq/series?u=paul&p=pass&time_precision=s", data, c)
 
-	time.Sleep(time.Second)
+	self.serverProcesses[0].WaitForServerToSync()
 
 	self.serverProcesses[0].QueryAsRoot("test_cq", "select mean(c1) from s3 group by time(5s) into d3.mean;", false, c)
 	self.serverProcesses[0].QueryAsRoot("test_cq", "select count(c2) from s3 group by time(5s) into d3.count;", false, c)
+
+	// wait for replication
+	self.serverProcesses[0].WaitForServerToSync()
+	// wait for the query to run
+	time.Sleep(time.Second)
 
 	collection := self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
 	series := collection.GetSeries("continuous queries", c)
@@ -1376,6 +668,11 @@ func (self *ServerSuite) TestContinuousQueryInterpolation(c *C) {
   ]`
 
 	self.serverProcesses[0].Post("/db/test_cq/series?u=paul&p=pass", data, c)
+
+	// wait for replication
+	self.serverProcesses[0].WaitForServerToSync()
+	// wait for the query to run
+	time.Sleep(time.Second)
 
 	collection := self.serverProcesses[0].QueryAsRoot("test_cq", "list continuous queries;", false, c)
 	series := collection.GetSeries("continuous queries", c)
@@ -1449,7 +746,9 @@ func (self *ServerSuite) TestCreateAndGetShards(c *C) {
 	}`, startSeconds, endSeconds, 2, 3)
 	resp := self.serverProcesses[0].Post("/cluster/shards?u=root&p=root", data, c)
 	c.Assert(resp.StatusCode, Equals, http.StatusAccepted)
-	time.Sleep(time.Second)
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToSync()
+	}
 	for _, s := range self.serverProcesses {
 		body := s.Get("/cluster/shards?u=root&p=root", c)
 		res := make(map[string]interface{})
@@ -1491,7 +790,12 @@ func (self *ServerSuite) TestDropShard(c *C) {
 	// now write some data to ensure that the local files get created
 	t := (time.Now().Unix() + secondsOffset) * 1000
 	data = fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_drop_shard", "columns": ["value", "time"]}]`, t)
-	self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
+	resp = self.serverProcesses[0].Post("/db/test_rep/series?u=paul&p=pass", data, c)
+	c.Assert(resp.StatusCode, Equals, http.StatusOK)
+
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToSync()
+	}
 
 	// and find the shard id
 	body := self.serverProcesses[0].Get("/cluster/shards?u=root&p=root", c)
@@ -1520,7 +824,9 @@ func (self *ServerSuite) TestDropShard(c *C) {
 
 	resp = self.serverProcesses[0].Delete(fmt.Sprintf("/cluster/shards/%d?u=root&p=root", shardId), data, c)
 	c.Assert(resp.StatusCode, Equals, http.StatusAccepted)
-	time.Sleep(time.Second)
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToSync()
+	}
 
 	for _, s := range self.serverProcesses {
 		body := s.Get("/cluster/shards?u=root&p=root", c)
@@ -1555,7 +861,9 @@ func (self *ServerSuite) TestDropShard(c *C) {
 
 	resp = self.serverProcesses[0].Delete(fmt.Sprintf("/cluster/shards/%d?u=root&p=root", shardId), data, c)
 	c.Assert(resp.StatusCode, Equals, http.StatusAccepted)
-	time.Sleep(time.Second)
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToSync()
+	}
 
 	for _, s := range self.serverProcesses {
 		body := s.Get("/cluster/shards?u=root&p=root", c)
@@ -1606,11 +914,18 @@ func (self *ServerSuite) TestContinuousQueryWithMixedGroupByOperations(c *C) {
   ]`)
 
 	self.serverProcesses[0].Post("/db/test_cq/series?u=paul&p=pass&time_precision=s", data, c)
-
+	// wait for the data to get written
+	self.serverProcesses[0].WaitForServerToSync()
+	// wait for the query to run
 	time.Sleep(time.Second)
 
 	self.serverProcesses[0].QueryAsRoot("test_cq", "select mean(reqtime), url from cqtest group by time(10s), url into cqtest.10s", false, c)
 	defer self.serverProcesses[0].QueryAsRoot("test_cq", "drop continuous query 1;", false, c)
+
+	// wait for the continuous query to run
+	time.Sleep(time.Second)
+	// wait for the continuous queries to propagate
+	self.serverProcesses[0].WaitForServerToSync()
 
 	collection := self.serverProcesses[0].QueryAsRoot("test_cq", "select * from cqtest.10s", false, c)
 	series := collection.GetSeries("cqtest.10s", c)
@@ -1624,7 +939,7 @@ func (self *ServerSuite) TestContinuousQueryWithMixedGroupByOperations(c *C) {
 }
 
 // fix for #305: https://github.com/influxdb/influxdb/issues/305
-func (self *ServerSuite) TestShardIdUniquenessAfterRestart(c *C) {
+func (self *ServerSuite) TestShardIdUniquenessAfterReStart(c *C) {
 	server := self.serverProcesses[0]
 	t := (time.Now().Unix() + 86400*720) * 1000
 	data := fmt.Sprintf(`[{"points": [[2, %d]], "name": "test_shard_id_uniqueness", "columns": ["value", "time"]}]`, t)
@@ -1650,10 +965,13 @@ func (self *ServerSuite) TestShardIdUniquenessAfterRestart(c *C) {
 	for _, s := range self.serverProcesses {
 		s.Stop()
 	}
-	time.Sleep(time.Second * 2)
+
 	for _, s := range self.serverProcesses {
 		s.Start()
-		time.Sleep(time.Second)
+	}
+
+	for _, s := range self.serverProcesses {
+		s.WaitForServerToStart()
 	}
 
 	server = self.serverProcesses[0]
