+++ src/integration/test_config3.toml
@@ -3,11 +3,11 @@
 # If hostname (on the OS) doesn't return a name that can be resolved by the other
 # systems in the cluster, you'll have to set the hostname to an IP or something
 # that can be resovled here.
-# hostname = ""
+hostname = "localhost"
 
 [logging]
 # logging level can be one of "debug", "info", "warn" or "error"
-level  = "debug"
+level  = "info"
 file   = "/tmp/influxdb/test/3/influxdb.log"
 
 # Configure the admin server
@@ -49,7 +49,9 @@ seed-servers = ["localhost:60501"]
 
 protobuf_port = 60512
 protobuf_timeout = "1200ms" # the write timeout on the protobuf conn any duration parseable by time.ParseDuration
-protobuf_heartbeat = "1s" # the heartbeat interval between the servers. must be parseable by time.ParseDuration
+protobuf_heartbeat = "100ms" # the heartbeat interval between the servers. must be parseable by time.ParseDuration
+protobuf_min_backoff = "100ms" # the minimum backoff after a failed heartbeat attempt
+protobuf_max_backoff = "100ms" # the maxmimum backoff after a failed heartbeat attempt
 
 # How many write requests to potentially buffer in memory per server. If the buffer gets filled then writes
 # will still be logged and once the server has caught up (or come back online) the writes
